delete modified

+++ src/PrivateRoute/MyArtCraftList.jsx
@@ -2,6 +2,8 @@ import { useContext, useEffect, useState } from "react";
 import { FaStar } from "react-icons/fa";
 import { AuthContext } from "../AuthProvider/AuthProvider";
 import { Link } from "react-router-dom";
+import Swal from 'sweetalert2';
+
 
 const MyArtCraftList = () => {
     const { user } = useContext(AuthContext);
@@ -21,16 +23,34 @@ const MyArtCraftList = () => {
 
     //delete item
     const handleDeleteItem = (id) => {
-        fetch(`http://localhost:5000/delete/${id}`, {
-            method: 'DELETE',
-
-        })
-        .then((res) => res.json())
-        .then((data) => {
-            if(data.deletedCount > 0){
-                setControl(!control)
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+          }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/delete/${id}`, {
+                    method: 'DELETE',
+        
+                })
+                .then((res) => res.json())
+                .then((data) => {
+                    if(data.deletedCount > 0){
+                        Swal.fire({
+                            title: "Deleted!",
+                            text: "Your file has been deleted.",
+                            icon: "success"
+                          });
+                        setControl(!control)
+                    }
+                })
             }
-        })
+          });
+        
     }
 
     return (
